// src/app/core/models/contact-excel.model.ts
export class ContactExcel {
  cedula: string;
  nombres: string;
  apellidos: string;
  email: string;
  celular: string;
  direccion: string;
  estado: string;
  tipo_contacto: string;
  solicito_acompañamiento: string;
  fecha_registro: string;
  nit: string;
  nombre_empresa: string;
  departamento_empresa: string;
  ciudad_empresa: string;
  direccion_empresa: string;
  email_empresa: string;
  telefono_empresa: string;
  celular_empresa: string;
  sitio_web: string;
}